# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Gym Application API
  version: 1.0.0
  description: API for managing gym trainers, trainees and their workouts.
  contact:
    name: Aman Nazarkulov
    email: aman_nazarkulov@epam.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/auth/change-password:
    put:
      operationId: changePassword
      summary: Change user password
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/UserCredentialsResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token